# perf: Improve Table Query Performance