10k row dataset, 10kb per row.

- (offset 200, limit 100)
  - Prod: Elapsed: 1.439s Read: 33,970 rows (225.50 MB)
  - _Master: Elapsed: 5.691s Read: 4,481,567 rows (18.02 GB)_
  - **PR: Elapsed: 0.098s Read: 34,531 rows (225.53 MB)**
- Filtered by Digest
  - Prod: Elapsed: 0.163s Read: 31,321 rows (225.39 MB)
  - _Master: Elapsed: 7.913s Read: 4,481,939 rows (18.02 GB)_
  - **PR: Elapsed: 0.052s Read: 24,222 rows (20.48 MB)**